+++ src/main/java/io/github/mjaroslav/sharedjava/reflect/UnableToFindGenericTypeException.java
@@ -1,0 +1,9 @@
+package io.github.mjaroslav.sharedjava.reflect;
+
+import org.jetbrains.annotations.NotNull;
+
+public class UnableToFindGenericTypeException extends RuntimeException {
+    public UnableToFindGenericTypeException(@NotNull Throwable cause) {
+        super(cause);
+    }
+}