+++ src/test/java/io/github/mjaroslav/sharedjava/reflect/EnumDummy.java
@@ -1,0 +1,5 @@
+package io.github.mjaroslav.sharedjava.reflect;
+
+public enum EnumDummy {
+    FIRST, SECOND, THIRD
+}